Update ug.js
there are many wrongs in this language file . For example,  translate "italic" to "Italian", “outdent” and "indent" is Increase the width and decrease the width, but I really don't know what this means by translation ，some of them are not translated . I've removed some unnecessary variable definitions. Also, I referred to the “UyghurSoft Dictionary”. In the end, I would like to thank @Ansar-786 , @blfrg and all Quasar Framework teams for their selfless dedication.